tiny bit of refactoring
