+++ src/components/country-list/country-list.funcs.js
@@ -1,4 +1,4 @@
-import { NOFLYLIST } from '../../config';
+import { NOFLYLIST } from '../../config/config';
 
 export const removeNoFly = (countries) => {
     const newCountries = countries.filter(country => !NOFLYLIST.includes(country.code));
