+++ src/config/config.js
@@ -1,4 +1,4 @@
-export const countryRoute = 'https://api.myjson.com/bins/16i597';
+export const apiEndPoint = 'https://api.myjson.com/bins/16i597';
 export const NOFLYLIST = [
     'TD', 'KP', 'VE', 'IR',
     'LY', 'SO', 'SY', 'YE',
